refactor(cli): run through managed execution manager
