+++ README.md
@@ -342,10 +342,13 @@ acquired by a failed start attempt. Ordinary `start` does not revive a terminal
 managed execution. CLI `restart` now uses a durable two-phase operation that
 terminates the old runtime before advancing the generation, recovers ambiguous
 kill/start responses from backend evidence, and rebinds local resources without
-duplicating ownership. CLI `run` and the Rust SDK still need migration. The production HCL
-service, encrypted credentials, generation-fenced route leases, wildcard TLS
-gateway, envd data plane, and real official-client Sandbox suite also remain
-open gates.
+duplicating ownership. CLI `run` now reserves and starts through the same
+manager, freezes image-defined health and stop defaults into the durable
+request, and leaves network, volume, rootfs, stop, and auto-remove ownership to
+the managed backend. The Rust SDK still needs migration. The production
+HCL service, encrypted credentials, generation-fenced route leases, wildcard
+TLS gateway, envd data plane, and real official-client Sandbox suite also
+remain open gates.
 
 The server, native Python/TypeScript packages, and unchanged-official-client
 black-box suites follow the phased design in
