+++ docs/e2b-compatible-sdk-design.md
@@ -1,8 +1,8 @@
 # E2B Protocol Compatibility and SDK Design
 
 Status: **Phase 1 complete; Phase 2 in progress (slices 1 through 3 complete;
-slice 4 runtime path and staged CLI create/start/restart complete, remaining
-callers pending)**
+slice 4 runtime path and CLI create/start/restart/run complete, Rust SDK
+pending)**
 
 Implementation evidence starts in [`compat/e2b/`](../compat/e2b/README.md).
 The pinned contract manifest intentionally reports `full_compatibility=false`;
@@ -22,7 +22,7 @@ unversioned claim.
 | Pinned contract | Vendored control, envd, volume-content, Process, Filesystem, MCP, public-export, and package artifacts with generated digests | Keep the manifest pinned and regenerate it only through reviewed upstream updates |
 | Lifecycle protocol | Owner-scoped create, connect, get, list, timeout, and kill routes; unchanged pinned Python sync/async, TypeScript, and Code Interpreter clients pass against the Rust fixture server | Run the same unchanged clients through the production service and a real Sandbox execution |
 | Durable control state | SQLite WAL migrations, strict record validation, compare-and-swap transitions, generation-fenced expiry claims, reaping, and startup reconciliation | Wire the repository and supervisor into the production service process and exercise restart and host-reboot recovery end to end |
-| Runtime lifecycle | Canonical managed-execution store, two-stage backend-neutral `LocalExecutionManager`, and production VM/Sandbox backend; CLI `create`, first `start`, and explicit two-phase `restart` use generation fencing with caller-policy parity, idempotent resource preparation, failure rollback, and operation-ID recovery tests; an A3S OS smoke test proves reservation-only create, creation reconciliation, real `crun` start, explicit pause rejection, kill, and cleanup without MicroVM fallback | Validate managed restart on A3S OS, then migrate CLI run and the Rust SDK and add the remaining caller parity tests |
+| Runtime lifecycle | Canonical managed-execution store, two-stage backend-neutral `LocalExecutionManager`, and production VM/Sandbox backend; CLI `create`, first `start`, `run`, and explicit two-phase `restart` use generation fencing with caller-policy parity, idempotent resource preparation, failure rollback, and operation-ID recovery tests; A3S OS smoke tests prove reservation-only create, creation reconciliation, real `crun` start, managed restart, explicit pause rejection, kill, and cleanup without MicroVM fallback | Validate managed CLI `run` on A3S OS, migrate the Rust SDK, and add its caller parity tests |
 | Credentials and routing | Injected verifier, token, cursor, and template interfaces isolate protocol logic from infrastructure | Add production credential hashing, token encryption and rotation, generation-fenced route leases, validated wildcard/direct routing, and the TLS data-plane gateway |
 | Commands and SDK surface | Pinned Process/Filesystem descriptors and Python/TypeScript public-export inventories prevent unreviewed drift | Implement envd HTTP, ConnectRPC, PTY, signed URLs, Code Interpreter/MCP streams, the remaining public control surface, and native convenience packages |
 
@@ -607,12 +607,11 @@ must not assume that the single-host limit is part of the upstream contract.
 ### Dependency direction
 
 The runtime now owns a canonical managed-execution store, a backend-neutral
-`LocalExecutionManager`, and a production VM/Sandbox backend. CLI `create`
-reserves through that manager, while CLI `start`/`run` still own separate boot
-paths and the Rust SDK intentionally omits those operations. The compatibility
-service must not work around the remaining caller gap by spawning `a3s-box`,
-importing CLI modules, or editing `boxes.json`. Phase 2 completes this
-dependency direction:
+`LocalExecutionManager`, and a production VM/Sandbox backend. CLI
+`create`/`start`/`restart`/`run` use that manager, while the Rust SDK
+intentionally omits those lifecycle operations. The compatibility service must
+not work around the remaining SDK gap by spawning `a3s-box`, importing CLI
+modules, or editing `boxes.json`. Phase 2 completes this dependency direction:
 
 ```text
 a3s-box-core
@@ -785,7 +784,8 @@ Phase 2 is delivered as small, immediately merged changes:
    `ExecutionManager`; add the production backend and prove its real Sandbox
    lifecycle; switch CLI create to the same reservation path; switch CLI
    start/restart/run and the Rust SDK to the same implementation with behavior
-   parity tests. Create, start, and restart are complete; run and the SDK remain.
+   parity tests. CLI create, start, restart, and run are complete; the SDK
+   remains.
 5. Add the production HCL-configured service binary, credential and token
    providers, generation-fenced route leases, and TLS data-plane gateway. Pull
    each merge commit on an A3S OS server and run the unmodified official clients
@@ -825,7 +825,8 @@ generation and requires a new operation ID for any later restart. Graceful-stop
 timeout is part of the restart intent, so a retry cannot silently change it.
 Named-volume and network ownership is released and rebound once, while
 execution-owned anonymous volumes remain available to the replacement
-generation; a terminal kill still removes them.
+generation. Retained terminal stops preserve those anonymous volumes for a
+later restart; auto-remove terminal kills remove them.
 
 The production VM/Sandbox backend is also complete for this slice. It owns live
 runtime handles, reconstructs MicroVM processes with PID identity fencing,
@@ -846,10 +847,18 @@ managed request, including config-only values such as DNS and persistent
 filesystem policy. Named-volume bookkeeping remains attached only after the
 durable reservation succeeds and rolls the reservation back on failure.
 
-Slice 4 remains incomplete until CLI run and the Rust SDK call the same manager
-with behavior parity tests. The existing Rust SDK uses the canonical
-record store for management operations but still has a separate local
-lifecycle model and does not expose create/start/run.
+CLI `run` now submits the complete caller policy to the same manager, starts
+under generation fencing, and reloads the canonical record before foreground
+or detached handling. It resolves image health and stop defaults before
+reservation, reuses the cache during backend start, and delegates network,
+volume, rootfs, stop, and auto-remove ownership to the managed backend. Caller
+parity tests cover isolation, DNS, environment, security, limits, TEE/sidecar,
+logs, health, stop policy, shared memory, persistence, and resource metadata.
+
+Slice 4 remains incomplete until the Rust SDK calls the same manager with
+behavior parity tests. The existing Rust SDK uses the canonical record store
+for management operations but still has a separate local lifecycle model and
+does not expose create/start/run.
 
 Each slice must pass its focused tests and repository CI before merge. The
 Phase 2 gate remains closed until slice 5 composes the durable repository,
